Added a view for recent measurements
