+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/ui/components/spl/SoundLevelMeterView.kt
@@ -8,7 +8,6 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -17,7 +16,6 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import org.jetbrains.compose.resources.stringResource
@@ -72,48 +70,7 @@ fun SoundLevelMeterView(
                 }
 
                 if (viewModel.showMinMaxSPL) {
-                    Row(
-                        Modifier.align(Alignment.Top),
-                    ) {
-                        listOf(
-                            LeqMetricViewModel(
-                                label = stringResource(viewModel.minDbALabel),
-                                value = currentLeqMetrics?.min,
-                            ),
-                            LeqMetricViewModel(
-                                label = stringResource(viewModel.avgDbALabel),
-                                value = currentLeqMetrics?.average,
-                            ),
-                            LeqMetricViewModel(
-                                label = stringResource(viewModel.maxDbALabel),
-                                value = currentLeqMetrics?.max,
-                            ),
-                        ).forEach { metric ->
-                            Column(
-                                horizontalAlignment = Alignment.Start,
-                                modifier = Modifier.width(56.dp),
-                            ) {
-                                Text(
-                                    text = metric.label,
-                                    style = MaterialTheme.typography.labelLarge
-                                )
-                                val value = metric.value
-                                Text(
-                                    text = if (value != null && value in VU_METER_DB_MIN..VU_METER_DB_MAX) {
-                                        value.toString()
-                                    } else "-",
-                                    style = MaterialTheme.typography.bodyMedium.copy(
-                                        fontWeight = FontWeight.Bold,
-                                        fontSize = 20.sp,
-                                        textAlign = TextAlign.Start,
-                                        color = metric.value?.let {
-                                            NoiseLevelColorRamp.getColorForSPLValue(it)
-                                        } ?: MaterialTheme.colorScheme.onSurface,
-                                    )
-                                )
-                            }
-                        }
-                    }
+                    LAeqMetricsView(metrics = currentLeqMetrics)
                 }
 
                 if (viewModel.showPlayPauseButton) {
@@ -133,9 +90,3 @@ fun SoundLevelMeterView(
         }
     }
 }
-
-
-private data class LeqMetricViewModel(
-    val label: String,
-    val value: Double?,
-)
